Extract RegisterValue and EvaluatedTuple into the register::value module;
